Added code for Lending service and modification for Book and patron services

+++ src/main/java/com/airtribe/LibraryManagement/repository/PatronRepository.java
@@ -3,5 +3,5 @@ package com.airtribe.LibraryManagement.repository;
 import com.airtribe.LibraryManagement.entity.Patron;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PatronRepository extends JpaRepository<Patron, Integer> {
+public interface PatronRepository extends JpaRepository<Patron, Long> {
 }
